pyproject.toml
[project]
name = "medbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Radhakrishnan A",email = "work.connectwithkrishna@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.26,<0.4.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "chromadb (>=1.0.13,<2.0.0)",
    "openai (>=1.91.0,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "langchain-community (>=0.3.26,<0.4.0)",
    "langchain-huggingface (>=0.3.0,<0.4.0)",
    "ipywidgets (>=8.1.7,<9.0.0)",
    "langchain-openai (>=0.3.25,<0.4.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pinecone-client (>=6.0.0,<7.0.0)",
    "langchain-core (>=0.3.66,<0.4.0)",
    "typing-extensions (>=4.14.0,<5.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "fpdf (>=1.7.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "medbot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
